Refactor canister_sig_util.
