+++ demos/vc_issuer/Cargo.toml
@@ -10,13 +10,14 @@ edition = "2021"
 # local dependencies
 canister_sig_util = { path = "../../src/canister_sig_util" }
 internet_identity_interface = { path = "../../src/internet_identity_interface" }
+vc_util = { path = "../../src/vc_util" }
 # ic dependencies
 candid = "0.9"
 ic-cdk = "0.10"
 ic-cdk-macros = "0.7"
 ic-certified-map = "0.4"
 # vc dependencies
-identity_core = {git = "https://github.com/frederikrothenberger/identity.rs.git", branch = "frederik/wasm-test", default-features = false}
+identity_core = { git = "https://github.com/frederikrothenberger/identity.rs.git", branch = "frederik/wasm-test", default-features = false}
 identity_credential = {git = "https://github.com/frederikrothenberger/identity.rs.git", branch = "frederik/wasm-test", default-features = false, features = ["credential"]}
 identity_jose = { git = "https://github.com/frederikrothenberger/identity.rs.git", branch = "frederik/wasm-test", default-features = false, features = ["iccs"]}
 # identity_core = {path = "../../../identity.rs/identity_core", default-features = false}
@@ -33,6 +34,5 @@ sha2 = "^0.10" # set bound to match ic-certified-map bound
 
 [dev-dependencies]
 ic-test-state-machine-client = "3"
-candid = { version = "0.9", features = ["parser"] }
 canister_tests = { path = "../../src/canister_tests" }
 lazy_static = "1.4"
