+++ src/internet_identity/Cargo.toml
@@ -6,6 +6,8 @@ edition = "2021"
 [dependencies]
 canister_sig_util = { path = "../../src/canister_sig_util" }
 internet_identity_interface = { path = "../internet_identity_interface" }
+vc_util = { path = "../../src/vc_util" }
+
 
 hex = "0.4"
 include_dir = "0.7"
